Decouple terrain and related file
