+++ src/GasGiant.h
@@ -7,7 +7,6 @@
 #include "BaseSphere.h"
 #include "GasGiantJobs.h"
 #include "JobQueue.h"
-#include "Random.h"
 #include "graphics/Material.h"
 #include "terrain/Terrain.h"
 #include "vector3.h"
