+++ src/GasGiantJobs.h
@@ -7,7 +7,6 @@
 #include <SDL_stdinc.h>
 
 #include "JobQueue.h"
-#include "Random.h"
 #include "graphics/Material.h"
 #include "graphics/opengl/GenGasGiantColourMaterial.h"
 #include "graphics/VertexBuffer.h"
