+++ src/terrain/TerrainFeature.cpp
@@ -1,6 +1,7 @@
 // Copyright © 2008-2019 Pioneer Developers. See AUTHORS.txt for details
 // Licensed under the terms of the GPL v3. See licenses/GPL-3.txt
 
+#include "FracDef.h"
 #include "TerrainFeature.h"
 #include "TerrainNoise.h"
